backend/requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-celery-results==2.5.1
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
kombu==5.3.2
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.9
